+++ tests/unit/test_denormalise_ons_data.py
@@ -5,7 +5,7 @@ import warnings
 from pyspark.sql import SparkSession
 
 import jobs.denormalise_ons_data as job
-from utils.ind_cqc_column_names.ons_columns import (
+from utils.column_names.raw_data_files.ons_columns import (
     OnsPostcodeDirectoryColumns as ColNames,
 )
